+++ packages/web/src/app/search/components/filterPanel/index.tsx
@@ -45,7 +45,6 @@ export const FilterPanel = ({
             "Language",
             matches,
             (key) => {
-                // @todo: Get language icons
                 return {
                     key,
                     displayName: key,
